requirements.txt
pip==21.1.2
pyserial==3.5
setuptools==57.0.0
wheel==0.36.2
